refactor(conf_loader): separate parsing functions into utils.lua
